Updating fabs derivations to have their own file and be split out into functions for ease of access
